Add ability to pass alternate swupd cert
Swupd now enforces valid certificates when installing content. Because
of this, when installing content that was not generated with the
certificate of the currently running system swupd will fail. To work
around this swupd added the '-C' option to pass an alternate
certificate. Support this use case in ister by adding the --cert-file
option which will take a path to the alternate certificate file.